Add instability with trail, behaviours
Added a more ellaborate version of the Instability sketch.
This has attractor, repeller behaviours, besides following.
Added findNearest and other methods, and a grid based optimization to
find the nearest particle trails around the particle trail of interest.
Set the canvas to PGraphics and draw an partially opaque black rectangle
on top to create the trail effects.
Moved the particle, grid functionality to their own files and remember
that in processing, the files in the root of the sketch are considerd to
be **in the same scope**!
Added a color palette class with utilities to create random RGB
palettes, or complementary, and triadic palettes, with a random offset
applied to hue, saturation and brigthness, as well as utils to return
them as hex, and to convert them from RGB to HSV and vice-versa. Note
that HSB is an alias for HSV in processing.